LayoutTests/imported/w3c/web-platform-tests/css/css-properties-values-api/animation/custom-property-animation-transform-none.tentative-expected.txt

CONSOLE MESSAGE: SyntaxError: The given initial value does not parse for the given syntax.

Harness Error (FAIL), message = SyntaxError: The given initial value does not parse for the given syntax.
